models/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
clip @ git+https://github.com/openai/CLIP.git
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=0.24.0
scipy>=1.7.0
matplotlib>=3.4.0
tqdm>=4.60.0
Pillow>=8.0.0 